requirements.txt
python-dotenv
pyannote.audio[separation]==3.3.2
loguru
scipy
torchaudio
